+++ lib/facebook_audience_network.dart
@@ -30,9 +30,11 @@ class FacebookAudienceNetwork {
   ///
   /// [testingId] can be obtained by running the app once without the testingId.
   /// Check the log to obtain the [testingId] for your device.
-  static Future<bool?> init({String? testingId}) async {
+  static Future<bool?> init(
+      {String? testingId, bool iOSAdvertiserTrackingEnabled = false}) async {
     Map<String, String?> initValues = {
       "testingId": testingId,
+      "iOSAdvertiserTrackingEnabled": iOSAdvertiserTrackingEnabled.toString(),
     };
 
     try {